Split getRecordingInfo into two properly-named files Fixes #365
