+++ src/containers/RecordingView.tsx
@@ -4,8 +4,8 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 import { RouteComponentProps, withRouter } from 'react-router';
 import { WorkspaceInfo } from '../AppContainer';
 import RecordingInfoView from '../components/RecordingInfoView';
-import { useRecordingInfo } from '../extensions/common/getRecordingInfo';
 import { createCalculationPool } from '../extensions/common/hither';
+import { useRecordingInfo } from '../extensions/common/useRecordingInfo';
 import { filterPlugins, Plugins, RecordingSelectionAction } from '../extensions/extensionInterface';
 import sortByPriority from '../extensions/sortByPriority';
 import { RootAction, RootState } from '../reducers';
