+++ src/containers/RecordingsTable.tsx
@@ -2,7 +2,7 @@ import { CircularProgress } from '@material-ui/core';
 import React, { FunctionComponent, useCallback, useMemo } from 'react';
 import { WorkspaceInfo } from '../AppContainer';
 import NiceTable from '../components/NiceTable';
-import { useRecordingInfos } from '../extensions/common/getRecordingInfo';
+import { useRecordingInfos } from '../extensions/common/useRecordingInfo';
 import { Recording, RecordingInfo } from '../reducers/recordings';
 import { Sorting, SortingInfo } from '../reducers/sortings';
 import { WorkspaceRouteDispatch } from './WorkspaceView';
