+++ src/containers/SortingUnitView.tsx
@@ -6,8 +6,8 @@ import { Link } from 'react-router-dom';
 import sizeMe, { SizeMeProps } from 'react-sizeme';
 import { WorkspaceInfo } from '../AppContainer';
 import SimilarUnit from '../components/SimilarUnit';
-import { useSortingInfo } from '../extensions/common/getRecordingInfo';
 import { CalculationPool, createCalculationPool, HitherContext } from '../extensions/common/hither';
+import { useSortingInfo } from '../extensions/common/useSortingInfo';
 import IndividualUnit from '../extensions/devel/IndividualUnits/IndividualUnit';
 import { filterPlugins, Plugins } from '../extensions/extensionInterface';
 import { ExtensionsConfig } from '../extensions/reducers';
