+++ src/containers/SortingView.tsx
@@ -4,8 +4,9 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 import { Link, RouteComponentProps, withRouter } from 'react-router-dom';
 import { setExternalSortingUnitMetrics } from '../actions';
 import { WorkspaceInfo } from '../AppContainer';
-import { useRecordingInfo, useSortingInfo } from '../extensions/common/getRecordingInfo';
 import { createCalculationPool, HitherContext } from '../extensions/common/hither';
+import { useRecordingInfo } from '../extensions/common/useRecordingInfo';
+import { useSortingInfo } from '../extensions/common/useSortingInfo';
 import { filterPlugins, Plugins, SortingCurationAction, SortingSelection, sortingSelectionReducer } from '../extensions/extensionInterface';
 import { getPathQuery } from '../kachery';
 import { RootAction, RootState } from '../reducers';
