+++ src/containers/WorkspaceRecordingView.tsx
@@ -3,8 +3,8 @@ import React, { FunctionComponent, useEffect, useReducer } from 'react';
 import { WorkspaceInfo } from '../AppContainer';
 import RecordingInfoView from '../components/RecordingInfoView';
 import SortingsView from '../components/SortingsView';
-import { useRecordingInfo } from '../extensions/common/getRecordingInfo';
 import { createCalculationPool } from '../extensions/common/hither';
+import { useRecordingInfo } from '../extensions/common/useRecordingInfo';
 import { Plugins, RecordingSelection, recordingSelectionReducer } from '../extensions/extensionInterface';
 import sortByPriority from '../extensions/sortByPriority';
 import { Recording } from '../reducers/recordings';
