+++ src/extensions/averagewaveforms/AverageWaveformsView/AverageWaveformsView.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent, useCallback, useEffect, useMemo, useState } from 'react';
 import { FaArrowDown, FaArrowUp } from 'react-icons/fa';
-import { useRecordingInfo } from '../../common/getRecordingInfo';
 import SortingUnitPlotGrid from '../../common/SortingUnitPlotGrid';
 import Splitter from '../../common/Splitter';
 import { ActionItem, DividerItem } from '../../common/Toolbars';
+import { useRecordingInfo } from '../../common/useRecordingInfo';
 import { SortingViewProps } from "../../extensionInterface";
 import AverageWaveformsToolbar from './AverageWaveformsToolbar';
 import AverageWaveformView from './AverageWaveformView';
