+++ src/extensions/devel/IndividualUnits/IndividualUnits.tsx
@@ -1,8 +1,8 @@
 import { Button, Grid } from '@material-ui/core';
 import React, { FunctionComponent, useState } from 'react';
 import sizeMe, { SizeMeProps } from 'react-sizeme';
-import { useSortingInfo } from '../../common/getRecordingInfo';
 import { createCalculationPool } from '../../common/hither';
+import { useSortingInfo } from '../../common/useSortingInfo';
 import { SortingViewProps } from '../../extensionInterface';
 import IndividualUnit from './IndividualUnit';
