+++ src/extensions/unitstable/Units/Units.tsx
@@ -1,8 +1,8 @@
 
 import { Button, Paper } from '@material-ui/core';
 import React, { useCallback, useContext, useEffect, useReducer, useState } from 'react';
-import { useSortingInfo } from '../../common/getRecordingInfo';
 import { HitherContext } from '../../common/hither';
+import { useSortingInfo } from '../../common/useSortingInfo';
 import { Recording, SortingUnitMetricPlugin, SortingViewProps } from '../../extensionInterface';
 import sortByPriority from '../../sortByPriority';
 import UnitsTable from './UnitsTable';
